Default to port 5060 when responding over UDP where the port is unspecified
This occurs when the Via header is not receiver tagged, and thus
follows Section 5 of RFC 3263 where the default port for the protocol
should be utilise if not specified in the sent-by.

This remains unnoticed by the tests due to the NkSIP UAC always
including the port in the 'Via' sent-by field.