Use appropriate URIs for From, To, Route, and Record-Route headers
In To, From, and Contact headers, a conforming endpoint should
consider parameters in a URI without angle brackets to be header
parameters.

Route and Record-Route header are specified as being `name-addr'
rather than `name-addr / addr-spec`.  As such these should _at least_
be wrapped with angle brackets.

Potentially there is room to make the client more user friendly in
this respect, but for now, correctness is probably more important.