Prevent monitor signal triggering an infinite loop in nksip_sipapp_srv
When an nksip_sipapp_srv process receives a 'DOWN' message from a
process that is not in the server's dictionary, then it will attempt
to forward the message to the underlying module.  This fails because
the pattern always matches.

A mod_handle_info/2 function provides the common operation for passing
a message to the passing on to the userspace SipApp.