Move nksip application to root path
It was somewhat difficult to include this project as a dependency with
the application located in a subdirectory.  This change is intended to
make the path structure somewhat more consistent with other Erlang
projects.

The only major quirk in the change is the need to specify
'../nksip/ebin' in vm.args.  This is necessary because of the way that
code:priv_dir/1 operates outside of the standard release structure.