Add callback for SipApps to indicate support for SIP extensions
This callback is utilise when the `Require' header is included in a
request.  By default, only extensions supported by NkSIP itself are
marked as supported.  Otherwise, a SipApp can indicate support by
returning a list of extensions.

Integration with the `make_supported' option is perhaps limited at
this stage.

See RFC 3261, 8.1.1.9 for more information.