Fix proxy's dialog behaviour
With the new code, allowing for different dialog's id for uac and uac, the dialog management at proxies is broken.
A proxy behaves in the same request as an uac and a uas, and the dialog id must be the same.
The proxy originates the request as a UAC. If a later in-dialog request comes from the other party, it is used the first dialog id instead of the new one.

A `dialog_id` field has been added to the #sipmsg{} record.
All tests should be passing again.